tools/opm/pyproject.toml
[project]
name = "opm"
version = "0.1.0"
description = "circle package manager"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "huggingface-hub>=1.2.1",
    "pyyaml>=6.0.3",
]
